Refactoring and fixing code: Refactored routes, fixed card navigation
