+++ app/utils/styles/Fonts.js
@@ -1,3 +1,3 @@
 export const Fonts = {
-    MontSerrat: 'Dosis-SemiBold.ttf'
+    MontSerrat: 'Montserrat.ttf'
 }
